requirements.txt
flasgger==0.9.7.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.37

